api delete user

+++ src/main/java/com/java/shinhan/team_fund_manage/repository/MemberRepository.java
@@ -4,6 +4,8 @@ import com.java.shinhan.team_fund_manage.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
 }
